Refactor toasts store
