+++ src/store/index.js
@@ -7,7 +7,6 @@ import i18nPlugin from './plugins/i18n'
 import router from './plugins/router'
 import loadingProgressReporter from './plugins/loadingProgressReporter'
 import dependentState from './plugins/dependentState'
-import toastsPlugin from './plugins/toasts'
 
 // Alphabetical
 import about from './modules/about'
@@ -81,7 +80,6 @@ export default new Vuex.Store({
     loadingProgressReporter,
     dependentState,
     fcmPlugin,
-    toastsPlugin,
   ],
   strict: debug,
 })
